requirements.txt
# W&B Ecosystem
wandb>=0.19.0
weave>=0.51.0

# Mistral
mistralai>=1.5.0

# OpenPipe ART (W&B Serverless Training)
art-e>=0.8.0

# Data handling
datasets>=3.0.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
httpx>=0.27.0
